+++ CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Fixed
 - Fixed the parsing of in alternation expressions `match` rules, i.e.
   `match a|b foo - 26`
-- Fix the parsing of `swapdd` operands
+- Fix the parsing of the `swapdd` operands
+- Fix the parsing of the `grouping` braille tuple
 
 ## [0.2.6] - 2026-01-27